Attempt to fix addon
I'm definitely no coder, and have zero experience with LUA and Ashita libraries but this does appear to work now using these changes. I swapped it to using /ms instead of /servo and then various changes to work with v3 Ashita.

I can only seem to get this to work while outside of combat, though. I'm really after finding a way to be able to send a dualbox my main's subtarget id while in combat, but if I'm engaged in combat then my dualbox doesn't seem to be getting it's target.